+++ src/Util/AbstractIO.hs
@@ -24,6 +24,8 @@ import Tezos.Common.Json (TezosInt64)
 import Tezos.Crypto
 import Tezos.V005.Micheline (Expression)
 
+import qualified Lorentz.Contracts.Multisig.Generic as MSig
+
 import Client.Error
 import Client.Types
 import Client.Env
@@ -51,11 +53,14 @@ class (Monad m) => HasCmdLine m where
 
 -- Interaction with tezos node via RPC
 class (HasTezosClient m, HasConfig m, Monad m, MonadThrow m) => HasTezosRpc m where
-  runTransactions :: (NicePackedValue param) => Address -> [EntrypointParam param] -> m ()
+  runTransactions
+    :: (NicePackedValue param)
+    => Address -> [(EntrypointParam param, TezosInt64)] -> m ()
   getStorage :: Text -> m Expression
   getCounter :: Text -> m TezosInt64
   getFromBigMap :: Natural -> Text -> m (Either TzbtcClientError Expression)
   deployTzbtcContract :: OriginationParameters -> m ()
+  deployMultisigContract :: MSig.Storage -> Bool -> m ()
 
 -- Interaction with tezos client binary
 class (HasConfig m, HasEnv m, Monad m) => HasTezosClient m where